# backend/internal/api/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AeroXe Bee API",
    "description": "Distributed SMS gateway platform with delivery confidence scoring and AI-powered template generation.",
    "version": "1.0.0",
    "contact": {
      "name": "Aeroxe Enterprises Pvt. Ltd.",
      "email": "support@aeroxbee.com",
      "url": "https://aeroxbee.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://aeroxbee.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development"
    },
    {
      "url": "https://api.aeroxbee.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/v1/member/send": {
      "post": {
        "tags": [
          "Member"
        ],
        "summary": "Send SMS via member portal",
        "operationId": "memberSend",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "device_id",
                  "recipient",
                  "message"
                ],
                "properties": {
                  "device_id": {
                    "type": "string",
                    "description": "Must belong to the member and be online"
                  },
                  "recipient": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string",
                    "maxLength": 160
                  },
                  "message_type": {
                    "type": "string",
                    "enum": [
                      "otp",
                      "transactional",
                      "marketing"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Message queued"
          },
          "400": {
            "description": "Validation error or device offline"
          },
          "403": {
            "description": "Device does not belong to account"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from login endpoint."
      },
      "APIKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API key passed as Bearer token."
      },
      "AdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin/staff JWT token with admin claim."
      }
    },
...
  }
}